chore!: Remove legacy native ListViewBase controls
Hard-removes the legacy native-only Uno.UI.Controls.Legacy ListViewBase
family (ListView/GridView/HorizontalListView/HorizontalGridView and their
Android.Widget.ListView / UICollectionView backing) and the internal
GridExtensions fluent helper that supported them, plus all SamplesApp
samples that consumed them.

- Deletes src/Uno.UI/UI/Xaml/Controls/ListViewBase/Legacy (native only).
- Deletes the legacy Uno.UI.Controls.Legacy.GridExtensions helper.
- Deletes the 9 legacy ListView/GridView/HorizontalListView samples.
- Migrates the two live Skia consumers of the fluent helpers
  (Basic_Shapes sample, When_Shape runtime test) to a self-contained
  local GridExtensions, and drops the now-stale legacy usings/xmlns from
  Given_Control, Given_Grid and the xBind phase samples.
- Drops the matching Uno.UI.Controls.Legacy entries from PackageDiffIgnore.

ProgressRing/Legacy (a separate, in-scope control sharing the namespace)
is intentionally left untouched.
