+++ src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml/xBindTests/NoPhaseBinding_Large.xaml
@@ -3,7 +3,6 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:legacy="using:Uno.UI.Controls.Legacy"
     xmlns:local="using:Uno.UI.Samples.Content.UITests.Performance"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:not_win="http://uno.ui/not_win"
