--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/GridView/GridViewLargeLegacy.xaml
@@ -1,46 +1,0 @@
-﻿<UserControl
-    x:Class="Uno.UI.Samples.Content.UITests.GridView.GridViewLargeLegacy"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:android="http://uno.ui/android"
-    xmlns:controls="using:Uno.UI.Samples.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:ios="http://uno.ui/ios"
-    xmlns:legacy="using:Uno.UI.Controls.Legacy"
-    xmlns:local="Uno.UI.Samples.Content.UITests.GridView"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:not_win="http://uno.ui/not_win"
-    xmlns:u="using:nVentive.Umbrella.Views.Controls"
-    xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    d:DesignHeight="2000"
-    d:DesignWidth="400"
-    mc:Ignorable="d ios android not_win">
-
-    <controls:SampleControl SampleDescription="Legacy GridView with large number of items">
-        <controls:SampleControl.SampleContent>
-            <DataTemplate>
-                <win:TextBlock Text="No legacy GridView on Windows, nothing to see here" />
-                <not_win:Border>
-                    <legacy:GridView
-                        Width="450"
-                        Height="400"
-                        android:ItemMaxWidth="80"
-                        ItemsSource="{Binding SampleItemsLong}">
-                        <legacy:GridView.ItemTemplate>
-                            <DataTemplate>
-                                <Border
-                                    Width="75"
-                                    Height="30"
-                                    Background="Green"
-                                    BorderBrush="GreenYellow"
-                                    BorderThickness="2">
-                                    <TextBlock Foreground="Yellow" Text="{Binding}" />
-                                </Border>
-                            </DataTemplate>
-                        </legacy:GridView.ItemTemplate>
-                    </legacy:GridView>
-                </not_win:Border>
-            </DataTemplate>
-        </controls:SampleControl.SampleContent>
-    </controls:SampleControl>
-</UserControl>
