--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/HorizontalListViewImage.xaml
@@ -1,41 +1,0 @@
-﻿<UserControl
-    x:Class="SamplesApp.Windows_UI_Xaml_Controls.ListView.HorizontalListViewImage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:android="http://nventive.com/android#using:Uno.UI.Controls.Legacy"
-    xmlns:controls="using:Uno.UI.Samples.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:ios="http://nventive.com/ios"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:u="using:Uno.UI.Samples.Controls"
-    xmlns:uBehaviors="using:Uno.UI.Samples.Behaviors"
-    xmlns:wasm="http://nventive.com/wasm"
-    xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    d:DesignHeight="2000"
-    d:DesignWidth="400"
-    mc:Ignorable="d ios android wasm">
-
-    <controls:SampleControl SampleDescription="You might have to unload and reload the sample to replicate -&gt; Bug 16510:[Android] [Image] Timing issue makes image not visible in Horizontal List View">
-        <controls:SampleControl.SampleContent>
-            <DataTemplate>
-                <ios:TextBlock Text="HorizontalListView is only used on Android." />
-                <win:TextBlock Text="HorizontalListView is only used on Android." />
-                <wasm:TextBlock Text="HorizontalListView is only used on Android." />
-                <android:HorizontalListView
-                    Width="250"
-                    Height="50"
-                    ItemsSource="{Binding [SampleItems]}"
-                    SelectionMode="None">
-                    <android:HorizontalListView.ItemTemplate>
-                        <DataTemplate>
-                            <Image
-                                Width="100"
-                                Height="50"
-                                Source="{Binding Uri}" />
-                        </DataTemplate>
-                    </android:HorizontalListView.ItemTemplate>
-                </android:HorizontalListView>
-            </DataTemplate>
-        </controls:SampleControl.SampleContent>
-    </controls:SampleControl>
-</UserControl>
