--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/HorizontalListViewImage.xaml.cs
@@ -1,15 +1,0 @@
-using Microsoft.UI.Xaml.Controls;
-using SamplesApp.Windows_UI_Xaml_Controls.Models;
-using Uno.UI.Samples.Controls;
-
-namespace SamplesApp.Windows_UI_Xaml_Controls.ListView
-{
-	[Sample("ListView", Name = "HorizontalListViewImage", ViewModelType = typeof(ImageListViewViewModel))]
-	public sealed partial class HorizontalListViewImage : UserControl
-	{
-		public HorizontalListViewImage()
-		{
-			this.InitializeComponent();
-		}
-	}
-}
