--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/HorizontalListViewTest.xaml
@@ -1,39 +1,0 @@
-﻿<UserControl
-    x:Class="SamplesApp.Windows_UI_Xaml_Controls.ListView.HorizontalListViewTest"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:TODO="http://uno.ui/TODO"
-    xmlns:android="http://uno.ui/android#using:Uno.UI.Controls.Legacy"
-    xmlns:controls="using:Uno.UI.Samples.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:ios="http://uno.ui/ios"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:u="using:Uno.UI.Samples.Controls"
-    xmlns:uBehaviors="using:Uno.UI.Samples.Behaviors"
-    xmlns:wasm="http://uno.ui/wasm"
-    xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    mc:Ignorable="d ios android wasm TODO">
-
-    <controls:SampleControl SampleDescription="HorizontalListViewTest">
-        <controls:SampleControl.SampleContent>
-            <DataTemplate>
-                <ios:TextBlock Text="HorizontalListView is only used on Android." />
-                <win:TextBlock Text="HorizontalListView is only used on Android." />
-                <wasm:TextBlock Text="HorizontalListView is only used on Android." />
-                <android:HorizontalListView
-                    Height="50"
-                    uBehaviors:ListViewBaseCommand.Command="{Binding [DoSomething]}"
-                    IsItemClickEnabled="true"
-                    ItemsSource="{Binding [SampleItems]}"
-                    SelectedItem="{Binding [SelectedItem]}"
-                    SelectionMode="Single">
-                    <android:HorizontalListView.ItemTemplate>
-                        <DataTemplate>
-                            <TextBlock Height="20" Text="{Binding}" />
-                        </DataTemplate>
-                    </android:HorizontalListView.ItemTemplate>
-                </android:HorizontalListView>
-            </DataTemplate>
-        </controls:SampleControl.SampleContent>
-    </controls:SampleControl>
-</UserControl>
