--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/HorizontalListView_Padding.xaml
@@ -1,42 +1,0 @@
-﻿<UserControl
-    x:Class="SamplesApp.Windows_UI_Xaml_Controls.ListView.HorizontalListView_Padding"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:android="http://uno.ui/android#using:Uno.UI.Controls.Legacy"
-    xmlns:controls="using:Uno.UI.Samples.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:ios="http://uno.ui/ios"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:u="using:Uno.UI.Samples.Controls"
-    xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    d:DesignHeight="2000"
-    d:DesignWidth="400"
-    mc:Ignorable="d ios android">
-
-    <controls:SampleControl SampleDescription="Sample control for demonstrating Padding on HorizontalListView">
-        <controls:SampleControl.SampleContent>
-            <DataTemplate>
-                <Border Background="Red">
-                    <ios:TextBlock Text="HorizontalListView is only used on Android." />
-                    <win:TextBlock Text="HorizontalListView is only used on Android." />
-                    <android:HorizontalListView
-                        Width="250"
-                        Height="50"
-                        Padding="100,0,100,0"
-                        Background="Green"
-                        ItemsSource="{Binding [SampleItems]}"
-                        SelectionMode="None">
-                        <android:HorizontalListView.ItemTemplate>
-                            <DataTemplate>
-                                <Image
-                                    Width="100"
-                                    Height="50"
-                                    Source="{Binding Uri}" />
-                            </DataTemplate>
-                        </android:HorizontalListView.ItemTemplate>
-                    </android:HorizontalListView>
-                </Border>
-            </DataTemplate>
-        </controls:SampleControl.SampleContent>
-    </controls:SampleControl>
-</UserControl>
