--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/HorizontalListView_Padding.xaml.cs
@@ -1,15 +1,0 @@
-using Microsoft.UI.Xaml.Controls;
-using SamplesApp.Windows_UI_Xaml_Controls.Models;
-using Uno.UI.Samples.Controls;
-
-namespace SamplesApp.Windows_UI_Xaml_Controls.ListView
-{
-	[Sample("ListView", Name = "HorizontalListView_Padding", ViewModelType = typeof(ImageListViewViewModel))]
-	public sealed partial class HorizontalListView_Padding : UserControl
-	{
-		public HorizontalListView_Padding()
-		{
-			this.InitializeComponent();
-		}
-	}
-}
