--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/HorizontalListView_Templates.xaml
@@ -1,71 +1,0 @@
-﻿<UserControl
-    x:Class="SamplesApp.Windows_UI_Xaml_Controls.ListView.HorizontalListView_Templates"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:android="http://uno.ui/android#using:Uno.UI.Controls.Legacy"
-    xmlns:behaviors="using:Uno.UI.Samples.Behaviors"
-    xmlns:controls="using:Uno.UI.Samples.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:ios="http://uno.ui/ios"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:u="using:Uno.UI.Samples.Controls"
-    xmlns:uConverters="using:Uno.UI.Samples.Converters"
-    xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    d:DesignHeight="2000"
-    d:DesignWidth="400"
-    mc:Ignorable="d ios android">
-
-    <UserControl.Resources>
-        <uConverters:FromNullableBoolToVisibilityConverter x:Key="boolToVisibility" />
-    </UserControl.Resources>
-
-    <controls:SampleControl SampleDescription="Sample to test the refresh of templates with binded values">
-        <controls:SampleControl.SampleContent>
-            <DataTemplate>
-                <StackPanel>
-                    <ios:TextBlock Text="HorizontalListView is only used on Android." />
-                    <win:TextBlock Text="HorizontalListView is only used on Android." />
-
-                    <android:TextBlock Text="Horizontal List View" />
-                    <android:HorizontalListView
-                        Height="200"
-                        behaviors:ListViewBaseCommand.Command="{Binding [DoSomething]}"
-                        IsItemClickEnabled="true"
-                        ItemsSource="{Binding [SampleItems]}"
-                        SelectedItem="{Binding [SelectedItem]}"
-                        SelectionMode="Single">
-                        <android:HorizontalListView.ItemTemplate>
-                            <DataTemplate>
-                                <StackPanel>
-                                    <TextBlock Height="20" Text="{Binding [Item]}" />
-                                    <Button Command="{Binding [MakeVisible]}" Content="Toggle" />
-                                    <Button Content="Hidden" Visibility="{Binding [IsVisible], Converter={StaticResource boolToVisibility}}" />
-                                </StackPanel>
-                            </DataTemplate>
-                        </android:HorizontalListView.ItemTemplate>
-                    </android:HorizontalListView>
-
-                    <android:TextBlock Text="List View" />
-                    <android:ListView
-                        Height="200"
-                        u:ListViewBaseCommand.Command="{Binding [DoSomething]}"
-                        IsItemClickEnabled="true"
-                        ItemsSource="{Binding [SampleItems]}"
-                        SelectedItem="{Binding [SelectedItem]}"
-                        SelectionMode="Single">
-                        <ListView.ItemTemplate>
-                            <DataTemplate>
-                                <StackPanel>
-                                    <TextBlock Height="20" Text="{Binding [Item]}" />
-                                    <Button Command="{Binding [MakeVisible]}" Content="Toggle" />
-                                    <Button Content="Hidden" Visibility="{Binding [IsVisible], Converter={StaticResource boolToVisibility}}" />
-                                </StackPanel>
-                            </DataTemplate>
-                        </ListView.ItemTemplate>
-                    </android:ListView>
-
-                </StackPanel>
-            </DataTemplate>
-        </controls:SampleControl.SampleContent>
-    </controls:SampleControl>
-</UserControl>
