--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/HorizontalListView_Templates.xaml.cs
@@ -1,15 +1,0 @@
-using Uno.UI.Samples.Controls;
-using Microsoft.UI.Xaml.Controls;
-using Uno.UI.Samples.Presentation.SamplePages;
-
-namespace SamplesApp.Windows_UI_Xaml_Controls.ListView
-{
-	[Sample("ListView", Name = "HorizontalListView_Templates", ViewModelType = typeof(ListViewRefreshTemplateViewModel))]
-	public sealed partial class HorizontalListView_Templates : UserControl
-	{
-		public HorizontalListView_Templates()
-		{
-			this.InitializeComponent();
-		}
-	}
-}
