--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/ListViewGroupedLargeLegacy.xaml
@@ -1,65 +1,0 @@
-﻿<UserControl
-    x:Class="SamplesApp.Windows_UI_Xaml_Controls.ListView.ListViewGroupedLargeLegacy"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:android="http://uno.ui/android"
-    xmlns:controls="using:Uno.UI.Samples.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:ios="http://uno.ui/ios"
-    xmlns:legacy="using:Uno.UI.Controls.Legacy"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:not_win="http://uno.ui/not_win"
-    xmlns:u="using:Uno.UI.Samples.Controls"
-    xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    d:DesignHeight="2000"
-    d:DesignWidth="400"
-    mc:Ignorable="d ios android not_win">
-
-    <controls:SampleControl SampleDescription="Legacy ListView with a grouped source with many items">
-        <controls:SampleControl.SampleContent>
-            <DataTemplate>
-                <win:TextBlock Text="No Legacy ListView on Windows, sample is Uno-specific." />
-                <not_win:Border>
-                    <legacy:ListView
-                        Width="200"
-                        Height="450"
-                        ItemsSource="{Binding [GroupedTownsAsSource]}">
-                        <legacy:ListView.GroupStyle>
-                            <GroupStyle HidesIfEmpty="False">
-                                <GroupStyle.HeaderTemplate>
-                                    <DataTemplate>
-                                        <Border
-                                            Background="Blue"
-                                            BorderBrush="Gray"
-                                            BorderThickness="2">
-                                            <TextBlock
-                                                Width="50"
-                                                Height="40"
-                                                FontSize="32"
-                                                Foreground="Red"
-                                                Text="{Binding Key}" />
-                                        </Border>
-                                    </DataTemplate>
-                                </GroupStyle.HeaderTemplate>
-                            </GroupStyle>
-                        </legacy:ListView.GroupStyle>
-                        <legacy:ListView.ItemTemplate>
-                            <DataTemplate>
-                                <Border
-                                    Background="White"
-                                    BorderBrush="Gray"
-                                    BorderThickness="2">
-                                    <TextBlock
-                                        Width="150"
-                                        Height="20"
-                                        Foreground="Blue"
-                                        Text="{Binding}" />
-                                </Border>
-                            </DataTemplate>
-                        </legacy:ListView.ItemTemplate>
-                    </legacy:ListView>
-                </not_win:Border>
-            </DataTemplate>
-        </controls:SampleControl.SampleContent>
-    </controls:SampleControl>
-</UserControl>
