--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/ListViewGroupedLargeLegacy.xaml.cs
@@ -1,15 +1,0 @@
-using SamplesApp.Windows_UI_Xaml_Controls.Models;
-using Uno.UI.Samples.Controls;
-using Microsoft.UI.Xaml.Controls;
-
-namespace SamplesApp.Windows_UI_Xaml_Controls.ListView
-{
-	[Sample("ListView", Name = "ListViewGroupedLargeLegacy", ViewModelType = typeof(ListViewGroupedViewModel))]
-	public sealed partial class ListViewGroupedLargeLegacy : UserControl
-	{
-		public ListViewGroupedLargeLegacy()
-		{
-			this.InitializeComponent();
-		}
-	}
-}
