--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/ListViewLargeLegacy.xaml
@@ -1,43 +1,0 @@
-﻿<UserControl
-    x:Class="SamplesApp.Windows_UI_Xaml_Controls.ListView.ListViewLargeLegacy"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:android="http://uno.ui/android"
-    xmlns:controls="using:Uno.UI.Samples.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:ios="http://uno.ui/ios"
-    xmlns:legacy="using:Uno.UI.Controls.Legacy"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:not_win="http://uno.ui/not_win"
-    xmlns:u="using:Uno.UI.Samples.Controls"
-    xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    d:DesignHeight="2000"
-    d:DesignWidth="400"
-    mc:Ignorable="d ios android not_win">
-
-    <controls:SampleControl SampleDescription="Legacy ListView with large number of items">
-        <controls:SampleControl.SampleContent>
-            <DataTemplate>
-                <win:TextBlock Text="No legacy ListView in Windows" />
-                <not_win:Border>
-                    <legacy:ListView
-                        Width="125"
-                        Height="400"
-                        ItemsSource="{Binding SampleItemsLong}">
-                        <legacy:ListView.ItemTemplate>
-                            <DataTemplate>
-                                <Border
-                                    Height="25"
-                                    Background="Green"
-                                    BorderBrush="GreenYellow"
-                                    BorderThickness="2">
-                                    <TextBlock Foreground="Yellow" Text="{Binding}" />
-                                </Border>
-                            </DataTemplate>
-                        </legacy:ListView.ItemTemplate>
-                    </legacy:ListView>
-                </not_win:Border>
-            </DataTemplate>
-        </controls:SampleControl.SampleContent>
-    </controls:SampleControl>
-</UserControl>
