--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/ListViewLargeLegacy.xaml.cs
@@ -1,15 +1,0 @@
-using Uno.UI.Samples.Controls;
-using Microsoft.UI.Xaml.Controls;
-using SamplesApp.Windows_UI_Xaml_Controls.Models;
-
-namespace SamplesApp.Windows_UI_Xaml_Controls.ListView
-{
-	[Sample("ListView", Name = "ListViewLargeLegacy", ViewModelType = typeof(ListViewViewModel))]
-	public sealed partial class ListViewLargeLegacy : UserControl
-	{
-		public ListViewLargeLegacy()
-		{
-			this.InitializeComponent();
-		}
-	}
-}
