--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/ListViewLegacy.xaml
@@ -1,39 +1,0 @@
-﻿<UserControl
-    x:Class="SamplesApp.Windows_UI_Xaml_Controls.ListView.ListViewLegacy"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:android="http://uno.ui/android"
-    xmlns:controls="using:Uno.UI.Samples.Controls"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:ios="http://uno.ui/ios"
-    xmlns:legacy="using:Uno.UI.Controls.Legacy"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:not_win="http://uno.ui/not_win"
-    xmlns:u="using:Uno.UI.Samples.Controls"
-    xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    d:DesignHeight="2000"
-    d:DesignWidth="400"
-    mc:Ignorable="d ios android not_win">
-
-    <controls:SampleControl SampleDescription="Legacy version of ListView - deprecated">
-        <controls:SampleControl.SampleContent>
-            <DataTemplate>
-                <win:TextBlock Text="Legacy ListView is Uno-specific, no Windows sample" />
-                <not_win:Border>
-                    <legacy:ListView
-                        u:ListViewBaseCommand.Command="{Binding [DoSomething]}"
-                        IsItemClickEnabled="true"
-                        ItemsSource="{Binding [SampleItems]}"
-                        SelectedItem="{Binding [SelectedItem]}"
-                        SelectionMode="Single">
-                        <legacy:ListView.ItemTemplate>
-                            <DataTemplate>
-                                <TextBlock Height="20" Text="{Binding}" />
-                            </DataTemplate>
-                        </legacy:ListView.ItemTemplate>
-                    </legacy:ListView>
-                </not_win:Border>
-            </DataTemplate>
-        </controls:SampleControl.SampleContent>
-    </controls:SampleControl>
-</UserControl>
