--- src/SamplesApp/SamplesApp.Samples/Windows_UI_Xaml_Controls/ListView/ListViewLegacy.xaml.cs
@@ -1,15 +1,0 @@
-using Uno.UI.Samples.Controls;
-using Microsoft.UI.Xaml.Controls;
-using SamplesApp.Windows_UI_Xaml_Controls.Models;
-
-namespace SamplesApp.Windows_UI_Xaml_Controls.ListView
-{
-	[Sample("ListView", Name = "ListViewLegacy", ViewModelType = typeof(ListViewViewModel))]
-	public sealed partial class ListViewLegacy : UserControl
-	{
-		public ListViewLegacy()
-		{
-			this.InitializeComponent();
-		}
-	}
-}
