+++ src/Uno.UI.UnitTests/Windows_UI_XAML_Controls/GridTests/Given_Grid.cs
@@ -5,7 +5,6 @@ using Microsoft.UI.Xaml.Controls;
 using System.Linq;
 using Windows.Foundation;
 using AwesomeAssertions.Execution;
-using Uno.UI.Controls.Legacy;
 using Uno.UI.Tests.Windows_UI_XAML_Controls.GridTests.Controls;
 
 namespace Uno.UI.Tests.GridTests
