--- src/Uno.UI/UI/Xaml/Controls/ListViewBase/Legacy/GridViewSource.Legacy.UIKit.cs
@@ -1,30 +1,0 @@
-﻿using System;
-using System.Linq;
-using Uno.Extensions;
-using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Controls.Primitives;
-
-using Foundation;
-using UIKit;
-using CoreGraphics;
-
-namespace Uno.UI.Controls.Legacy
-{
-	public partial class GridViewSource : ListViewBaseSource
-	{
-		public GridViewSource(GridView owner = null) : base(owner)
-		{
-
-		}
-		protected override SelectorItem CreateSelectorItem()
-		{
-			return new GridViewItem() { ShouldHandlePressed = false };
-		}
-
-		protected override ListViewBaseHeaderItem CreateSectionHeaderItem()
-		{
-			return new GridViewHeaderItem();
-		}
-	}
-}
-
