--- src/Uno.UI/UI/Xaml/Controls/ListViewBase/Legacy/ListViewBaseScrollDirection.UIKit.cs
@@ -1,22 +1,0 @@
-﻿using Uno.Extensions;
-using Microsoft.UI.Xaml;
-using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-
-using Foundation;
-using UIKit;
-using CoreGraphics;
-
-namespace Microsoft.UI.Xaml.Controls
-{
-	public enum ListViewBaseScrollDirection
-	{
-		Vertical,
-		Horizontal,
-	}
-	//
-	// FIXME Merge with ListView and remove the ListViewTypeSelector in favor of using a generic "ListViewItem" and ItemTemplate for its content
-
-}
