--- src/Uno.UI/UI/Xaml/Controls/ListViewBase/Legacy/ListViewSource.Legacy.UIKit.cs
@@ -1,31 +1,0 @@
-﻿using System;
-using System.Linq;
-using Uno.Extensions;
-using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Controls.Primitives;
-using System.Collections.Generic;
-
-using Foundation;
-using UIKit;
-using CoreGraphics;
-
-namespace Uno.UI.Controls.Legacy
-{
-	public partial class ListViewSource : ListViewBaseSource
-	{
-		public ListViewSource(ListView owner) : base(owner)
-		{
-		}
-
-		protected override SelectorItem CreateSelectorItem()
-		{
-			return new ListViewItem() { ShouldHandlePressed = false };
-		}
-
-		protected override ListViewBaseHeaderItem CreateSectionHeaderItem()
-		{
-			return new ListViewHeaderItem();
-		}
-	}
-}
-
